usvg::parser::svgtree

Struct SvgNode

pub struct SvgNode<'a, 'input: 'a> {
    id: NodeId,
    doc: &'a Document<'input>,
    d: &'a NodeData,
}

An SVG node.

Fields
id: NodeIddoc: &'a Document<'input>d: &'a NodeData
Implementations

impl<'a, 'input: 'a> SvgNode<'a, 'input>

pub(crate) fn convert_length( &self, aid: AId, object_units: Units, state: &State<'_>, def: Length, ) -> f32

pub fn convert_user_length( &self, aid: AId, state: &State<'_>, def: Length, ) -> f32

pub fn parse_viewbox(&self) -> Option<NonZeroRect>

pub fn resolve_length(&self, aid: AId, state: &State<'_>, def: f32) -> f32

pub fn resolve_valid_length( &self, aid: AId, state: &State<'_>, def: f32, ) -> Option<NonZeroPositiveF32>

pub(crate) fn try_convert_length( &self, aid: AId, object_units: Units, state: &State<'_>, ) -> Option<f32>

pub fn has_valid_transform(&self, aid: AId) -> bool

pub fn is_visible_element(&self, opt: &Options<'_>) -> bool

impl SvgNode<'_, '_>

pub(crate) fn resolve_transform( &self, transform_aid: AId, state: &State<'_>, ) -> Transform

impl<'a, 'input: 'a> SvgNode<'a, 'input>

fn id(&self) -> NodeId

pub fn is_element(&self) -> bool

Checks if the current node is an element.

pub fn is_text(&self) -> bool

Checks if the current node is a text.

pub fn document(&self) -> &'a Document<'input>

Returns node’s document.

pub fn tag_name(&self) -> Option<EId>

Returns element’s tag name, unless the current node is text.

pub fn element_id(&self) -> &'a str

Returns element’s id attribute value.

Returns an empty string otherwise.

pub fn attribute<T: FromValue<'a, 'input>>(&self, aid: AId) -> Option<T>

Returns an attribute value.

pub fn try_attribute<T: FromValue<'a, 'input>>(&self, aid: AId) -> Option<T>

Returns an attribute value.

Same as SvgNode::attribute, but doesn’t show a warning.

fn node_attribute(&self, aid: AId) -> Option<SvgNode<'a, 'input>>

pub fn has_attribute(&self, aid: AId) -> bool

Checks if an attribute is present.

pub fn attributes(&self) -> &'a [Attribute<'input>]

Returns a list of all element’s attributes.

fn attribute_id(&self, aid: AId) -> Option<usize>

pub fn find_attribute<T: FromValue<'a, 'input>>(&self, aid: AId) -> Option<T>

Finds a [Node] that contains the required attribute.

For inheritable attributes walks over ancestors until a node with the specified attribute is found.

For non-inheritable attributes checks only the current node and the parent one. As per SVG spec.

fn find_attribute_impl(&self, aid: AId) -> Option<SvgNode<'a, 'input>>

pub fn text(&self) -> &'a str

Returns node’s text data.

For text nodes returns its content. For elements returns the first child node text.

pub fn parent(&self) -> Option<Self>

Returns a parent node.

pub fn parent_element(&self) -> Option<Self>

Returns the parent element.

pub fn next_sibling(&self) -> Option<Self>

Returns the next sibling.

pub fn first_child(&self) -> Option<Self>

Returns the first child.

pub fn first_element_child(&self) -> Option<Self>

Returns the first child element.

pub fn last_child(&self) -> Option<Self>

Returns the last child.

pub fn has_children(&self) -> bool

Checks if the node has child nodes.

pub fn ancestors(&self) -> Ancestors<'a, 'input>

Returns an iterator over ancestor nodes starting at this node.

pub fn children(&self) -> Children<'a, 'input>

Returns an iterator over children nodes.

fn traverse(&self) -> Traverse<'a, 'input>

Returns an iterator which traverses the subtree starting at this node.

pub fn descendants(&self) -> Descendants<'a, 'input>

Returns an iterator over this node and its descendants.

pub fn href_iter(&self) -> HrefIter<'a, 'input>

Returns an iterator over elements linked via xlink:href.