Crate usvg::rosvgtree

Represents an SVG document as a read-only tree.

rosvgtree is similar to roxmltree, and even uses it for parsing, but instead of producing an XML tree, it produces an SVG tree by post-processing XML, to make SVG parsing easier.

Modules

roxmltree

Represent an XML 1.0 document as a read-only tree.

Structs

Ancestors

An iterator over ancestor nodes.

Attribute

An attribute.

Children

An iterator over children nodes.

Descendants

A descendants iterator.

Document

An SVG tree container.

HrefIter

An iterator over xlink:href references.

Node

An SVG node.

Enums

AttributeId

An attribute ID.

ElementId

An element ID.

Error

A list of all possible errors.