usvg

Struct FontResolver

pub struct FontResolver<'a> {
    pub select_font: FontSelectionFn<'a>,
    pub select_fallback: FallbackSelectionFn<'a>,
}

A font resolver for <text> elements.

This type can be useful if you want to have an alternative font handling to the default one. By default, only fonts specified upfront in Options::fontdb will be used. This type allows you to load additional fonts on-demand and customize the font selection process.

Fields
select_font: FontSelectionFn<'a>

Resolver function that will be used when selecting a specific font for a generic Font specification.

select_fallback: FallbackSelectionFn<'a>

Resolver function that will be used when selecting a fallback font for a character.

Implementations

impl FontResolver<'_>

pub fn default_font_selector() -> FontSelectionFn<'static>

Creates a default font selection resolver.

The default implementation forwards to query on the font database specified in the Options.

pub fn default_fallback_selector() -> FallbackSelectionFn<'static>

Creates a default font fallback selection resolver.

The default implementation searches through the entire fontdb to find a font that has the correct style and supports the character.