usvg

Struct Options

pub struct Options<'a> {
    pub resources_dir: Option<PathBuf>,
    pub dpi: f32,
    pub font_family: String,
    pub font_size: f32,
    pub languages: Vec<String>,
    pub shape_rendering: ShapeRendering,
    pub text_rendering: TextRendering,
    pub image_rendering: ImageRendering,
    pub default_size: Size,
    pub image_href_resolver: ImageHrefResolver<'a>,
    pub font_resolver: FontResolver<'a>,
    pub fontdb: Arc<Database>,
}

Processing options.

Fields
resources_dir: Option<PathBuf>

Directory that will be used during relative paths resolving.

Expected to be the same as the directory that contains the SVG file, but can be set to any.

Default: None

dpi: f32

Target DPI.

Impacts units conversion.

Default: 96.0

font_family: String

A default font family.

Will be used when no font-family attribute is set in the SVG.

Default: Times New Roman

font_size: f32

A default font size.

Will be used when no font-size attribute is set in the SVG.

Default: 12

languages: Vec<String>

A list of languages.

Will be used to resolve a systemLanguage conditional attribute.

Format: en, en-US.

Default: [en]

shape_rendering: ShapeRendering

Specifies the default shape rendering method.

Will be used when an SVG element’s shape-rendering property is set to auto.

Default: GeometricPrecision

text_rendering: TextRendering

Specifies the default text rendering method.

Will be used when an SVG element’s text-rendering property is set to auto.

Default: OptimizeLegibility

image_rendering: ImageRendering

Specifies the default image rendering method.

Will be used when an SVG element’s image-rendering property is set to auto.

Default: OptimizeQuality

default_size: Size

Default viewport size to assume if there is no viewBox attribute and the width or height attributes are relative.

Default: (100, 100)

image_href_resolver: ImageHrefResolver<'a>

Specifies the way xlink:href in <image> elements should be handled.

Default: see type’s documentation for details

font_resolver: FontResolver<'a>

Specifies how fonts should be resolved and loaded.

fontdb: Arc<Database>

A database of fonts usable by text.

This is a base database. If a custom font_resolver is specified, additional fonts can be loaded during parsing. Those will be added to a copy of this database. The full database containing all fonts referenced in a Tree becomes available as Tree::fontdb after parsing. If no fonts were loaded dynamically, that database will be the same as this one.

Implementations

impl Options<'_>

pub fn get_abs_path(&self, rel_path: &Path) -> PathBuf

Converts a relative path into absolute relative to the SVG file itself.

If Options::resources_dir is not set, returns itself.

pub fn fontdb_mut(&mut self) -> &mut Database

Mutably acquires the database.

This clones the database if it is currently shared.