Struct usvg::PathData

pub struct PathData { /* private fields */ }

An SVG path data container.

All segments are in absolute coordinates.

Implementations

impl PathData

pub fn new() -> PathData

Creates a new path.

pub fn is_empty(&self) -> bool

Returns true if the path contains no segment.

pub fn len(&self) -> usize

Returns the number of segments in the path.

pub fn commands(&self) -> &[PathCommand]

Returns a slice of the path commands.

pub fn points(&self) -> &[f64]

Returns a slice of the path points.

pub fn clear(&mut self)

Clears the path.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the path as much as possible.

pub fn from_rect(rect: Rect) -> PathData

Creates a path from a rect.

pub fn push_move_to(&mut self, x: f64, y: f64)

Pushes a MoveTo segment to the path.

pub fn push_line_to(&mut self, x: f64, y: f64)

Pushes a LineTo segment to the path.

pub fn push_quad_to(&mut self, x1: f64, y1: f64, x: f64, y: f64)

Pushes a QuadTo segment to the path.

Will be converted into cubic curve.

pub fn push_curve_to(&mut self, x1: f64, y1: f64, x2: f64, y2: f64, x: f64, y: f64)

Pushes a CurveTo segment to the path.

pub fn push_arc_to( &mut self, rx: f64, ry: f64, x_axis_rotation: f64, large_arc: bool, sweep: bool, x: f64, y: f64 )

Pushes an ArcTo segment to the path.

Arc will be converted into cubic curves.

pub fn push_close_path(&mut self)

Pushes a ClosePath segment to the path.

pub fn push_rect(&mut self, rect: Rect)

Pushes a rect to the path.

pub fn push_path(&mut self, path: &PathData)

Pushes a path to the path.

pub fn bbox(&self) -> Option<PathBbox>

Calculates path’s bounding box.

This operation is expensive.

pub fn bbox_with_transform( &self, ts: Transform, stroke: Option<&Stroke> ) -> Option<PathBbox>

Calculates path’s bounding box with a specified transform.

This operation is expensive.

pub fn has_bbox(&self) -> bool

Checks that path has a bounding box.

This operation is expensive.

pub fn length(&self) -> f64

Calculates path’s length.

Length from the first segment to the first MoveTo, ClosePath or slice end.

This operation is expensive.

pub fn transform(&mut self, ts: Transform)

Applies the transform to the path.

pub fn transform_from(&mut self, offset: usize, ts: Transform)

Applies the transform to the path from the specified offset.

pub fn segments(&self) -> PathSegmentsIter<'_>

Returns an iterator over path segments.