usvg

Struct Tree

pub struct Tree { /* private fields */ }

A nodes tree container.

Implementations

impl Tree

pub fn from_data(data: &[u8], opt: &Options<'_>) -> Result<Self, Error>

Parses Tree from an SVG data.

Can contain an SVG string or a gzip compressed data.

pub fn from_str(text: &str, opt: &Options<'_>) -> Result<Self, Error>

Parses Tree from an SVG string.

pub fn from_xmltree( doc: &Document<'_>, opt: &Options<'_>, ) -> Result<Self, Error>

Parses Tree from roxmltree::Document.

impl Tree

pub fn size(&self) -> Size

Image size.

Size of an image that should be created to fit the SVG.

width and height in SVG.

pub fn root(&self) -> &Group

The root element of the SVG tree.

pub fn node_by_id(&self, id: &str) -> Option<&Node>

Returns a renderable node by ID.

If an empty ID is provided, than this method will always return None.

pub fn has_text_nodes(&self) -> bool

Checks if the current tree has any text nodes.

pub fn linear_gradients(&self) -> &[Arc<LinearGradient>]

Returns a list of all unique LinearGradients in the tree.

pub fn radial_gradients(&self) -> &[Arc<RadialGradient>]

Returns a list of all unique RadialGradients in the tree.

pub fn patterns(&self) -> &[Arc<Pattern>]

Returns a list of all unique Patterns in the tree.

pub fn clip_paths(&self) -> &[Arc<ClipPath>]

Returns a list of all unique ClipPaths in the tree.

pub fn masks(&self) -> &[Arc<Mask>]

Returns a list of all unique Masks in the tree.

pub fn filters(&self) -> &[Arc<Filter>]

Returns a list of all unique Filters in the tree.

pub fn fontdb(&self) -> &Arc<Database>

Returns the font database that applies to all text nodes in the tree.

impl Tree

pub fn to_string(&self, opt: &WriteOptions) -> String

Writes usvg::Tree back to SVG.