usvg::text

Module layout

Provides access to the layout of a text node.

Structs

ByteIndex 🔒

A read-only text index in bytes.

DecorationSpan 🔒

A text decoration span.

Glyph 🔒

A glyph.

GlyphCluster 🔒

GlyphClusters 🔒

An iterator over glyph clusters.

PathNormal 🔒

PositionedGlyph

A glyph that has already been positioned correctly.

ResolvedFont 🔒

Span

A span contains a number of layouted glyphs that share the same fill, stroke, paint order and visibility.

Traits

DatabaseExt 🔒

Functions

apply_length_adjust 🔒

apply_letter_spacing 🔒

Applies the letter-spacing property to a text chunk clusters.

apply_word_spacing 🔒

Applies the word-spacing property to a text chunk clusters.

apply_writing_mode 🔒

Rotates clusters according to Unicode Vertical_Orientation Property.

chunk_span_at 🔒

clusters_length 🔒

collect_decoration_spans 🔒

collect_normals 🔒

convert_decoration 🔒

convert_span 🔒

form_glyph_clusters 🔒

is_word_separator_characters 🔒

Checks that the selected character is a word separator.

layout_text 🔒

process_anchor 🔒

process_chunk 🔒

Converts a text chunk into a list of outlined clusters.

resolve_baseline 🔒

resolve_baseline_shift 🔒

resolve_clusters_positions 🔒

Resolves clusters positions.

resolve_clusters_positions_horizontal 🔒

resolve_clusters_positions_path 🔒

script_supports_letter_spacing 🔒

Checks that selected script supports letter spacing.

shape_text 🔒

Text shaping with font fallback.

shape_text_with_font 🔒

Converts a text into a list of glyph IDs.

span_contains 🔒

Type Aliases

FontsCache 🔒