usvg

Module tree

Re-exports

pub use strict_num;

pub use tiny_skia_path;

pub use self::geom::*;

pub use self::text::*;

Modules

filter

SVG filter types.

geom 🔒

text 🔒

Structs

BaseGradient

A generic gradient.

ClipPath

A clip-path element.

Color

A 8-bit RGB color.

Fill

A fill style.

Group

A group container.

Image

A raster image element.

LinearGradient

A linear gradient.

Mask

A mask element.

NonEmptyString 🔒

NonZeroF32

A non-zero f32.

NonZeroPositiveF32

An immutable, finite f32 that is known to be > 0.

NormalizedF32

An immutable, finite f32 in a 0..=1 range.

Path

A path element.

Pattern

A pattern element.

PositiveF32

An immutable, finite f32 that is known to be >= 0.

RadialGradient

A radial gradient.

Stop

Gradient’s stop element.

Stroke

A stroke style.

StrokeMiterlimit

A stroke-miterlimit value.

Tree

A nodes tree container.

Enums

BlendMode

A blending mode property.

ContextElement 🔒

FillRule

A fill rule.

ImageKind

An embedded image kind.

ImageRendering

An image rendering method.

LineCap

A line cap.

LineJoin

A line join.

MaskType

A mask type.

Node

Node’s kind.

Paint

A paint style.

PaintOrder

Representation of the paint-order property.

ShapeRendering

A shape rendering method.

SpreadMethod

A spread method.

TextRendering

A text rendering method.

Units 🔒

Visibility 🔒

A visibility property.

Traits

ApproxEqUlps

ApproxEqUlps is a trait for approximate equality comparisons. The associated type Flt is a floating point type which implements Ulps, and is required so that this trait can be implemented for compound types (e.g. vectors), not just for the floats themselves.

Functions

has_text_nodes 🔒

loop_over_paint_servers 🔒

node_by_id 🔒

Type Aliases

Opacity

An alias to NormalizedF32.

StopOffset

An alias to NormalizedF32.

StrokeWidth

An alias to NonZeroPositiveF32.