usvg::tree

Struct Group

pub struct Group {

    pub(crate) id: String,
    pub(crate) transform: Transform,
    pub(crate) abs_transform: Transform,
    pub(crate) opacity: Opacity,
    pub(crate) blend_mode: BlendMode,
    pub(crate) isolate: bool,
    pub(crate) clip_path: Option<Arc<ClipPath>>,
    pub(crate) is_context_element: bool,
    pub(crate) mask: Option<Arc<Mask>>,
    pub(crate) filters: Vec<Arc<Filter>>,
    pub(crate) bounding_box: Rect,
    pub(crate) abs_bounding_box: Rect,
    pub(crate) stroke_bounding_box: Rect,
    pub(crate) abs_stroke_bounding_box: Rect,
    pub(crate) layer_bounding_box: NonZeroRect,
    pub(crate) abs_layer_bounding_box: NonZeroRect,
    pub(crate) children: Vec<Node>,

}

A group container.

The preprocessor will remove all groups that don’t impact rendering. Those that left is just an indicator that a new canvas should be created.

g element in SVG.

Fields
id: Stringtransform: Transformabs_transform: Transformopacity: Opacityblend_mode: BlendModeisolate: boolclip_path: Option<Arc<ClipPath>>is_context_element: bool

Whether the group is a context element (i.e. a use node)

mask: Option<Arc<Mask>>filters: Vec<Arc<Filter>>bounding_box: Rectabs_bounding_box: Rectstroke_bounding_box: Rectabs_stroke_bounding_box: Rectlayer_bounding_box: NonZeroRectabs_layer_bounding_box: NonZeroRectchildren: Vec<Node>
Implementations

impl Group

pub(crate) fn empty() -> Self

pub fn id(&self) -> &str

Element’s ID.

Taken from the SVG itself. Isn’t automatically generated. Can be empty.

pub fn transform(&self) -> Transform

Element’s transform.

This is a relative transform. The one that is set via the transform attribute in SVG.

pub fn abs_transform(&self) -> Transform

Element’s absolute transform.

Contains all ancestors transforms including group’s transform.

Note that subroots, like clipPaths, masks and patterns, have their own root transform, which isn’t affected by the node that references this subroot.

pub fn opacity(&self) -> Opacity

Group opacity.

After the group is rendered we should combine it with a parent group using the specified opacity.

pub fn blend_mode(&self) -> BlendMode

Group blend mode.

mix-blend-mode in SVG.

pub fn isolate(&self) -> bool

Group isolation.

isolation in SVG.

pub fn clip_path(&self) -> Option<&ClipPath>

Element’s clip path.

pub fn mask(&self) -> Option<&Mask>

Element’s mask.

pub fn filters(&self) -> &[Arc<Filter>]

Element’s filters.

pub fn bounding_box(&self) -> Rect

Element’s object bounding box.

objectBoundingBox in SVG terms. Meaning it doesn’t affected by parent transforms.

Can be set to None in case of an empty group.

pub fn abs_bounding_box(&self) -> Rect

Element’s bounding box in canvas coordinates.

userSpaceOnUse in SVG terms.

pub fn stroke_bounding_box(&self) -> Rect

Element’s object bounding box including stroke.

Similar to bounding_box, but includes stroke.

pub fn abs_stroke_bounding_box(&self) -> Rect

Element’s bounding box including stroke in user coordinates.

Similar to abs_bounding_box, but includes stroke.

pub fn layer_bounding_box(&self) -> NonZeroRect

Element’s “layer” bounding box in object units.

Conceptually, this is stroke_bounding_box expanded and/or clipped by filters_bounding_box, but also including all the children. This is the bounding box resvg will later use to allocate layers/pixmaps during isolated groups rendering.

Only groups have it, because only groups can have filters. For other nodes layer bounding box is the same as stroke bounding box.

Unlike other bounding boxes, cannot have zero size.

Returns 0x0x1x1 for empty groups.

pub fn abs_layer_bounding_box(&self) -> NonZeroRect

Element’s “layer” bounding box in canvas units.

pub fn children(&self) -> &[Node]

Group’s children.

pub fn should_isolate(&self) -> bool

Checks if this group should be isolated during rendering.

pub fn has_children(&self) -> bool

Returns true if the group has any children.

pub fn filters_bounding_box(&self) -> Option<NonZeroRect>

Calculates a node’s filter bounding box.

Filters with objectBoundingBox and missing or zero bounding_box would be ignored.

Note that a filter region can act like a clipping rectangle, therefore this function can produce a bounding box smaller than bounding_box.

Returns None when then group has no filters.

This function is very fast, that’s why we do not store this bbox as a Group field.

fn subroots(&self, f: &mut dyn FnMut(&Group))

impl Group

pub(crate) fn collect_clip_paths(&self, clip_paths: &mut Vec<Arc<ClipPath>>)

pub(crate) fn collect_masks(&self, masks: &mut Vec<Arc<Mask>>)

pub(crate) fn collect_filters(&self, filters: &mut Vec<Arc<Filter>>)

pub(crate) fn calculate_object_bbox(&mut self) -> Option<NonZeroRect>

pub(crate) fn calculate_bounding_boxes(&mut self) -> Option<()>