Crate usvg

usvg (micro SVG) is an SVG parser that tries to solve most of SVG complexity.

SVG is notoriously hard to parse. usvg presents a layer between an XML library and a potential SVG rendering library. It will parse an input SVG into a strongly-typed tree structure were all the elements, attributes, references and other SVG features are already resolved and presented in a simplest way possible. So a caller doesn’t have to worry about most of the issues related to SVG parsing and can focus just on the rendering part.

Features

All supported attributes are resolved. No need to worry about inheritable, implicit and default attributes

CSS will be applied

Only simple paths

Basic shapes (like rect and circle) will be converted into paths

Paths contain only absolute MoveTo, LineTo, CurveTo and ClosePath segments. ArcTo, implicit and relative segments will be converted

use will be resolved and replaced with the reference content

Nested svg will be resolved

Invalid, malformed elements will be removed

Relative length units (mm, em, etc.) will be converted into pixels/points

External images will be loaded

Internal, base64 images will be decoded

All references (like #elem and url(#elem)) will be resolved

switch will be resolved

Text elements, which are probably the hardest part of SVG, will be completely resolved. This includes all the attributes resolving, whitespaces preprocessing (xml:space), text chunks and spans resolving

Markers will be converted into regular elements. No need to place them manually

All filters are supported. Including filter functions, like filter="contrast(50%)"

Recursive elements will be detected an removed

Limitations

Unsupported SVG features will be ignored

CSS support is minimal

Only static SVG features, e.g. no a, view, cursor, script, no events and no animations

Text elements must be converted into paths before writing to SVG.

Modules

filter

SVG filter types.

fontdb

fontdb is a simple, in-memory font database with CSS-like queries.

roxmltree

Represent an XML 1.0 document as a read-only tree.

strict_num

A collection of bounded numeric types.

tiny_skia_path

A tiny-skia Bezier path implementation.

utils

Some useful utilities.

Structs

AspectRatio

Representation of the preserveAspectRatio attribute.

BBox

A bounding box calculator.

BaseGradient

A generic gradient.

ClipPath

A clip-path element.

Color

A 8-bit RGB color.

Fill

A fill style.

Font

Text font properties.

Group

A group container.

Image

A raster image element.

ImageHrefResolver

An xlink:href resolver for <image> elements.

LinearGradient

A linear gradient.

Mask

A mask element.

NonZeroF32

A non-zero f32.

NonZeroPositiveF32

An immutable, finite f32 that is known to be > 0.

NonZeroRect

A rectangle defined by left, top, right and bottom edges.

NormalizedF32

An immutable, finite f32 in a 0..=1 range.

Options

Processing options.

Path

A path element.

Pattern

A pattern element.

PositiveF32

An immutable, finite f32 that is known to be >= 0.

RadialGradient

A radial gradient.

Rect

A rectangle defined by left, top, right and bottom edges.

Size

A size.

Stop

Gradient’s stop element.

Stroke

A stroke style.

StrokeMiterlimit

A stroke-miterlimit value.

Text

A text element.

TextChunk

A text chunk.

TextDecoration

A text span decoration.

TextDecorationStyle

A text span decoration style.

TextPath

A path used by text-on-path.

TextSpan

A text style span.

Transform

An affine transformation matrix.

Tree

A nodes tree container.

ViewBox

View box.

XmlOptions

XML writing options.

Enums

AId

An attribute ID.

Align

Representation of the align value of the preserveAspectRatio attribute.

AlignmentBaseline

An alignment baseline property.

BaselineShift

A baseline shift property.

BlendMode

A blending mode property.

DominantBaseline

A dominant baseline property.

EId

An element ID.

Error

List of all errors.

FillRule

A fill rule.

FontStretch

A font stretch property.

FontStyle

A font style property.

ImageKind

An embedded image kind.

ImageRendering

An image rendering method.

LengthAdjust

A length adjust property.

LineCap

A line cap.

LineJoin

A line join.

MaskType

A mask type.

NodeKind

Node’s kind.

Paint

A paint style.

PaintOrder

Representation of the paint-order property.

ShapeRendering

A shape rendering method.

SpreadMethod

A spread method.

TextAnchor

A text chunk anchor property.

TextFlow

A text chunk flow property.

TextRendering

A text rendering method.

Units

An element units.

Visibility

A visibility property.

WritingMode

A writing mode.

Traits

ApproxEqUlps

ApproxEqUlps is a trait for approximate equality comparisons. The associated type Flt is a floating point type which implements Ulps, and is required so that this trait can be implemented for compound types (e.g. vectors), not just for the floats themselves.

ApproxZeroUlps

Approximate zero equality comparisons.

IsValidLength

Checks that the current number is > 0.

NodeExt

Additional Node methods.

TreeParsing

A trait to parse usvg_tree::Tree from various sources.

TreeTextToPath

A usvg::Tree extension trait.

TreeWriting

A trait to write usvg::Tree back to SVG.

Functions

decompress_svgz

Decompresses an SVGZ file.

Type Aliases

Node

Alias for rctree::Node<NodeKind>.

Opacity

An alias to NormalizedF32.

StopOffset

An alias to NormalizedF32.

StrokeWidth

An alias to NonZeroPositiveF32.