usvg::parser::svgtree

Struct Document

pub struct Document<'input> {
    nodes: Vec<NodeData>,
    attrs: Vec<Attribute<'input>>,
    links: HashMap<String, NodeId>,
}

An SVG tree container.

Contains only element and text nodes. Text nodes are present only inside the text element.

Fields
nodes: Vec<NodeData>attrs: Vec<Attribute<'input>>links: HashMap<String, NodeId>
Implementations

impl<'input> Document<'input>

pub fn parse_tree( xml: &Document<'input>, injected_stylesheet: Option<&'input str>, ) -> Result<Document<'input>, Error>

Parses a Document from a roxmltree::Document.

pub(crate) fn append(&mut self, parent_id: NodeId, kind: NodeKind) -> NodeId

fn append_attribute( &mut self, name: AId, value: StringStorage<'input>, important: bool, )

impl<'input> Document<'input>

pub fn root<'a>(&'a self) -> SvgNode<'a, 'input>

Returns the root node.

pub fn root_element<'a>(&'a self) -> SvgNode<'a, 'input>

Returns the root element.

pub fn descendants<'a>(&'a self) -> Descendants<'a, 'input>

Returns an iterator over document’s descendant nodes.

Shorthand for doc.root().descendants().

pub fn element_by_id<'a>(&'a self, id: &str) -> Option<SvgNode<'a, 'input>>

Returns an element by ID.

Unlike the Descendants iterator, this is just a HashMap lookup. Meaning it’s way faster.

fn get<'a>(&'a self, id: NodeId) -> SvgNode<'a, 'input>