Struct usvg::Tree

pub struct Tree {
    pub size: Size,
    pub view_box: ViewBox,
    pub root: Node<NodeKind>,
}

A nodes tree container.

Fields
size: Size

Image size.

Size of an image that should be created to fit the SVG.

width and height in SVG.

view_box: ViewBox

SVG viewbox.

Specifies which part of the SVG image should be rendered.

viewBox and preserveAspectRatio in SVG.

root: Node<NodeKind>

The root element of the SVG tree.

The root node is always Group.

Implementations

impl Tree

pub fn node_by_id(&self, id: &str) -> Option<Node<NodeKind>>

Returns renderable node by ID.

If an empty ID is provided, than this method will always return None. Even if tree has nodes with empty ID.

pub fn has_text_nodes(&self) -> bool

Checks if the current tree has any text nodes.

pub fn paint_servers<F>(&self, f: F)where F: FnMut(&Paint),

Calls a closure for each Paint in the tree.

Doesn’t guarantee to have unique paint servers. A caller must deduplicate them manually.

pub fn clip_paths<F>(&self, f: F)where F: FnMut(Rc<ClipPath>),

Calls a closure for each ClipPath in the tree.

Doesn’t guarantee to have unique clip paths. A caller must deduplicate them manually.

pub fn masks<F>(&self, f: F)where F: FnMut(Rc<Mask>),

Calls a closure for each Mask in the tree.

Doesn’t guarantee to have unique masks. A caller must deduplicate them manually.

pub fn filters<F>(&self, f: F)where F: FnMut(Rc<Filter>),

Calls a closure for each Filter in the tree.

Doesn’t guarantee to have unique filters. A caller must deduplicate them manually.