usvg::text

Module layout

Provides access to the layout of a text node.

Structs

ByteIndex ğŸ”’

A read-only text index in bytes.

DecorationSpan ğŸ”’

A text decoration span.

Glyph ğŸ”’

A glyph.

GlyphCluster ğŸ”’

GlyphClusters ğŸ”’

An iterator over glyph clusters.

PathNormal ğŸ”’

PositionedGlyph

A glyph that has already been positioned correctly.

ResolvedFont ğŸ”’

Span

A span contains a number of layouted glyphs that share the same fill, stroke, paint order and visibility.

Traits

DatabaseExt ğŸ”’

Functions

apply_length_adjust ğŸ”’

apply_letter_spacing ğŸ”’

Applies the letter-spacing property to a text chunk clusters.

apply_word_spacing ğŸ”’

Applies the word-spacing property to a text chunk clusters.

apply_writing_mode ğŸ”’

Rotates clusters according to Unicode Vertical_Orientation Property.

chunk_span_at ğŸ”’

clusters_length ğŸ”’

collect_decoration_spans ğŸ”’

collect_normals ğŸ”’

convert_decoration ğŸ”’

convert_span ğŸ”’

form_glyph_clusters ğŸ”’

is_word_separator_characters ğŸ”’

Checks that the selected character is a word separator.

layout_text ğŸ”’

process_anchor ğŸ”’

process_chunk ğŸ”’

Converts a text chunk into a list of outlined clusters.

resolve_baseline ğŸ”’

resolve_baseline_shift ğŸ”’

resolve_clusters_positions ğŸ”’

Resolves clusters positions.

resolve_clusters_positions_horizontal ğŸ”’

resolve_clusters_positions_path ğŸ”’

script_supports_letter_spacing ğŸ”’

Checks that selected script supports letter spacing.

shape_text ğŸ”’

Text shaping with font fallback.

shape_text_with_font ğŸ”’

Converts a text into a list of glyph IDs.

span_contains ğŸ”’

Type Aliases

FontsCache ğŸ”’